Added the logic to dynamically retrieve the model to be used during serialization and deserialization - common_models.py
