+++ container_service_extension/rde/converters/utils.py
@@ -1,25 +1,9 @@
 # container-service-extension
 # Copyright (c) 2021 VMware, Inc. All Rights Reserved.
 # SPDX-License-Identifier: BSD-2-Clause
-import abc
-
+from container_service_extension.rde.converters.abstract_convert_strategy import \
+    AbstractConvertStrategy
 from container_service_extension.rde.models.utils import get_model
-
-
-class AbstractConvertStrategy(abc.ABC):
-    def __init__(self):
-        pass
-
-    @abc.abstractmethod
-    def convert(self, data, src_rde_version: str, destn_rde_version: str):
-        """Convert the Cluster Entity from one RDE version to another.
-
-        :param  data: Input cluster entity
-        :param src_rde_version: Source RDE version
-        :param str destn_rde_version: Destination RDE version
-        :return data:
-        """
-        pass
 
 
 def convert(data: dict, destn_rde_version: str, src_rde_version: str = None,
