+++ container_service_extension/rde/models/abstractNativeEntity.py
@@ -1,0 +1,9 @@
+# container-service-extension
+# Copyright (c) 2021 VMware, Inc. All Rights Reserved.
+# SPDX-License-Identifier: BSD-2-Clause
+
+import abc
+
+
+class AbstractNativeEntity(abc.ABC):
+    pass
