+++ container_service_extension/rde/models/utils.py
@@ -1,3 +1,6 @@
+# container-service-extension
+# Copyright (c) 2021 VMware, Inc. All Rights Reserved.
+# SPDX-License-Identifier: BSD-2-Clause
 
 import semantic_version
